Flavor Details Document
This document is just a description of some potential
non-gameplay-changing features that possibly could be included in the
game, if possible. I'm mostly just uploading it so that you can look at
it and weigh in on how possible or not these might be, or if we'd have
to make early adjustments for it to work out. It's not a final list by
any means, nor are these necessarily things I'd horribly want badly in
the game. They're just prospects to look over.